requirements.txt
# ============================================================
# DocMind 依赖清单
# pip install -r requirements.txt
# ============================================================

# --- LangChain 生态 ---
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-huggingface>=0.0.3
langchain-chroma>=0.1.0
langchain-text-splitters>=0.2.0

# --- 向量数据库 ---
chromadb>=0.5.0

# --- Embedding & ML ---
sentence-transformers>=2.7.0
torch>=2.0.0

# --- 文档解析 ---
PyMuPDF>=1.24.0           # PDF 文字提取
python-docx>=1.0.0        # Word 文档解析
pdfplumber>=0.10.0        # PDF 表格提取（Week 2 进阶）

# --- Web 框架 ---
streamlit>=1.30.0         # 前端 UI
fastapi>=0.110.0          # 后端 API（Day 3-4）
uvicorn>=0.27.0           # ASGI 服务器
sse-starlette>=1.8.0      # SSE 流式推送

# --- 工具库 ---
PyYAML>=6.0               # 配置文件解析
loguru>=0.7.0             # 日志系统
rank-bm25>=0.2.2          # BM25 关键词检索（Week 2）
jieba>=0.42.1             # 中文分词（Week 2）
requests>=2.31.0          # API 请求（前后端分离）

# --- OpenAI 兼容 ---
openai>=1.12.0
